requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
reportlab==4.0.9
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9